rdx: Add cast avoid warning from gcc
Avoid a warning from gcc about casting an integer of different size to
a pointer.

+++ rdoff/rdx.c
@@ -51,7 +51,7 @@ int main(int argc, char **argv)
         exit(255);
     }
 
-    code = (main_fn) s->offset;
+    code = (main_fn)(size_t) s->offset;
 
     argv++, argc--;             /* remove 'rdx' from command line */
